+++ CHANGELOG.md
@@ -5,3 +5,5 @@
   - Updated to use SDC crypto library
   - Ensure integrity and version of library dependencies
   - Namespace config
+  - Add healthcheck endpoint
+